requirements.txt
python-telegram-bot==20.7
flask==3.0.0
playwright==1.40.0
python-dotenv==1.0.0
requests==2.31.0
urllib3<2.0.0
